Update#3
Correction of float formatting, adjustment for touch and scrolling

+++ Core/Inc/numpad.h
@@ -6,7 +6,7 @@
 #include "button.h"
 
 
-float number ;
+double number ;
 int digitNumber, fractionCount;
 uint8_t isFloat;
 void checkNumpadButtons();
